move value-related methods to an oracle; remove util scope slurping and resolve explicitly
