+++ src/tokenizer/lexer/token_lexers/rvalue.rb
@@ -3,7 +3,7 @@ module Tokenizer
     module TokenLexers
       # An rvalue is either a primitive, special identifier, or scoped variable.
       def rvalue?(chunk)
-        value?(chunk) || @current_scope.variable?(chunk)
+        Oracles::Value.value?(chunk) || @current_scope.variable?(chunk)
       end
 
       # TODO: (v1.1.0) Cannot assign keys / indices to themselves. (Fix at same time as process_attribute)
